+++ example/otel-collector/main.go
@@ -80,8 +80,7 @@ func initProvider() func() {
 	pusher.Start()
 
 	return func() {
-		ctx := context.Background()
-		handleErr(bsp.Shutdown(ctx), "failed to stop processor")
+		handleErr(tracerProvider.Shutdown(ctx), "failed to shutdown provider")
 		handleErr(exp.Shutdown(ctx), "failed to stop exporter")
 		pusher.Stop() // pushes any last exports to the receiver
 	}